♻️  macros: A big refactoring sweep
